+++ lib/tailwind_ui.dart
@@ -1,4 +1,5 @@
 library tailwind_ui;
 
 export 'src/application_ui/elements/avatars/circular_avatar.dart';
+export 'src/application_ui/elements/avatars/circular_avatar_with_notification.dart';
 export 'src/application_ui/elements/avatars/rounded_avatar.dart';